LayoutTests/platform/gtk/imported/w3c/web-platform-tests/shadow-dom/focus-navigation/focus-navigation-web-component-radio-expected.txt

OUT
A B
C D
OUT

FAIL Focus for web component input type elements should be bound by <form> inside shadow DOM assert_equals: expected "A" but got "start"
